+++ server/common/src/main/kotlin/su/plo/voice/server/audio/source/VoiceServerProximitySource.kt
@@ -10,6 +10,7 @@ import su.plo.voice.api.server.audio.source.AudioSender
 import su.plo.voice.api.server.audio.source.ServerProximitySource
 import su.plo.voice.api.server.event.audio.source.ServerSourceAudioPacketEvent
 import su.plo.voice.api.server.event.audio.source.ServerSourcePacketEvent
+import su.plo.voice.api.server.socket.UdpServerConnection
 import su.plo.voice.proto.data.audio.codec.CodecInfo
 import su.plo.voice.proto.data.audio.source.SourceInfo
 import su.plo.voice.proto.packets.Packet
@@ -34,8 +35,11 @@ abstract class VoiceServerProximitySource<S : SourceInfo>(
             field = value
         }
 
-    override fun sendAudioPacket(packet: SourceAudioPacket, distance: Short, activationInfo: PlayerActivationInfo?): Boolean {
-        // call event
+    override fun sendAudioPacket(
+        packet: SourceAudioPacket,
+        distance: Short,
+        activationInfo: PlayerActivationInfo?,
+    ): Boolean {
         val event = ServerSourceAudioPacketEvent(this, packet, distance, activationInfo)
         if (!voiceServer.eventBus.fire(event)) return false
         if (event.result == ServerSourceAudioPacketEvent.Result.HANDLED) return true
@@ -43,53 +47,28 @@ abstract class VoiceServerProximitySource<S : SourceInfo>(
         // update packet's source state
         packet.sourceState = state.get().toByte()
 
-        val listenersDistance = min(event.distance + voiceServer.config!!.voice().maxExtraAudioBroadcastDistance(), event.distance * DISTANCE_MULTIPLIER)
-
         // update source info on listeners if source is dirty
         if (dirty.compareAndSet(true, false)) {
             resolveSourceInfo()
-                .thenAccept { sendPacket(SourceInfoPacket(it), listenersDistance.toShort()) }
+                .thenAccept { sendPacket(SourceInfoPacket(it), event.distance) }
         }
 
-        val playerPosition = ServerPos3d()
-        val sourcePosition = position
-        val distanceSquared = (listenersDistance * listenersDistance).toDouble()
-
-        for (connection in voiceServer.udpConnectionManager.connections) {
-            if (notMatchFilters(connection.player)) continue
-
-            connection.player.instance.getServerPosition(playerPosition)
-            if (sourcePosition.world == playerPosition.world &&
-                sourcePosition.distanceSquared(playerPosition) <= distanceSquared
-            ) {
-                connection.sendPacket(packet)
-            }
+        getListeners(event.distance).forEach { connection ->
+            connection.sendPacket(packet)
         }
+
         return true
     }
 
     override fun sendPacket(packet: Packet<*>, distance: Short): Boolean {
-        // call event
         val event = ServerSourcePacketEvent(this, packet, distance)
         if (!voiceServer.eventBus.fire(event)) return false
         if (event.result == ServerSourcePacketEvent.Result.HANDLED) return true
 
-        val listenersDistance = min(event.distance + voiceServer.config!!.voice().maxExtraAudioBroadcastDistance(), event.distance * DISTANCE_MULTIPLIER)
-
-        val playerPosition = ServerPos3d()
-        val sourcePosition = position
-        val distanceSquared = (listenersDistance * listenersDistance).toDouble()
-
-        for (connection in voiceServer.udpConnectionManager.connections) {
-            if (notMatchFilters(connection.player)) continue
-
-            connection.player.instance.getServerPosition(playerPosition)
-            if (sourcePosition.world == playerPosition.world &&
-                sourcePosition.distanceSquared(playerPosition) <= distanceSquared
-            ) {
-                connection.player.sendPacket(packet)
-            }
+        getListeners(event.distance).forEach { connection ->
+            connection.player.sendPacket(packet)
         }
+
         return true
     }
 
@@ -102,6 +81,30 @@ abstract class VoiceServerProximitySource<S : SourceInfo>(
             { frame, sequenceNumber -> sendAudioFrame(frame, sequenceNumber, distanceProvider.get()) },
             { sequenceNumber -> sendAudioEnd(sequenceNumber, distanceProvider.get()) }
         )
+
+    override fun getListeners(distance: Short): Iterable<UdpServerConnection> {
+        val listenersDistance = min(
+            distance + voiceServer.config!!.voice().maxExtraAudioBroadcastDistance(),
+            distance * DISTANCE_MULTIPLIER,
+        )
+
+        val distanceSquared = (listenersDistance * listenersDistance).toDouble()
+
+        return Iterable {
+            val sourcePosition = position
+            val playerPosition = ServerPos3d()
+
+            voiceServer.udpConnectionManager.connections
+                .asSequence()
+                .filter { matchFilters(it.player) }
+                .filter { connection ->
+                    connection.player.instance.getServerPosition(playerPosition)
+
+                    sourcePosition.world == playerPosition.world && sourcePosition.distanceSquared(playerPosition) <= distanceSquared
+                }
+                .iterator()
+        }
+    }
 
     companion object {