+++ client/changelog.md
@@ -3,3 +3,4 @@
   - In other game modes, it was also moved slightly lower: https://i.imgur.com/CUpcIdc.png.
 
 ### Server
+- Added `BaseServerDirectSource#getListeners` and `ServerProximitySource#getListeners` to API to resolve current source's listeners.